+++ framework/Database/Seeder.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Framework\Database;
+
+abstract class Seeder implements SeederInterface
+{
+}