+++ framework/registerFn.php
@@ -15,7 +15,7 @@ function view(string $name, array $data = []): void
     if (count($data) > 0) {
         extract($data);
     }
-    require Path::join(__DIR__,  '..', 'resources', 'views', str_replace('.', '/', $name) . '.php');
+    require Path::join(__DIR__,  '..', 'resources', 'Views', str_replace('.', '/', $name) . '.php');
 }
 
 /** Render the given component */
@@ -24,7 +24,7 @@ function component(string $name, array $data = []): void
     if (count($data) > 0) {
         extract($data);
     }
-    require Path::join(__DIR__, '..', 'resources', 'views', 'components', str_replace('.', '/', $name) . '.php');
+    require Path::join(__DIR__, '..', 'resources', 'Views', 'Components', str_replace('.', '/', $name) . '.php');
 }
 
 /** Translate the given label into user language */