+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added message class to pass messages from controller to view
 - Added session class to set and get values in or from session
 - Added form and controller to change the password
+- Added database seeding system with runner and command create a new seeder
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2